+++ packages/spry_router/lib/spry_router.dart
@@ -1,6 +1,7 @@
 library spry.router;
 
 export 'src/param_middleware.dart';
+export 'src/handler_middleware_extension.dart';
 export 'src/request_params_extension.dart';
 export 'src/route.dart';
 export 'src/router.dart';