+++ packages/spry_router/CHANGELOG.md
@@ -1,3 +1,21 @@
+## 0.1.1
+
+### Param Middleware
+
+Param middleware now supports `use` method to add middleware to a specific param.
+
+```dart
+final ParamMiddleware middleware.use(otherMiddleware);
+```
+
+### Handler extension
+
+Spry handler supported `use` and `param` methods to add middleware and param middleware to a handler.
+
+```dart
+final Handler handler.use(middleware).param(paramMiddleware);
+```
+
 ## 0.1.0
 
 1. Update `spry` to `0.1.3`.